Add Quick Start tutorial for VSE
The tutorial is based on the cordova's quick start tutorial for the photo album.
We do not want the Telerik Platform to make changes in their code base at the moment and for this reason a new root vse specific metadata file (**vse-metadata.json**) is added which points only to the **vse-quick-start**. The Telerik Platform will just ignore the newly added files.